requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.21.1,<0.22.0
sqlalchemy>=2.0.9,<2.1.0
psycopg2-binary>=2.9.6,<2.10.0
alembic>=1.10.3,<1.11.0
pydantic>=1.10.7,<2.0.0
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
python-multipart>=0.0.6,<0.1.0
httpx>=0.24.0,<0.25.0
stripe>=5.4.0,<5.5.0
boto3>=1.26.131,<1.27.0
python-dotenv>=1.0.0,<1.1.0
bcrypt>=4.0.1,<4.1.0
email-validator>=2.0.0,<2.1.0
